Removed all references to sleep time;  I'm using awaits with promises to go between exercises after the voice pormpt ends. Added ability to pause timer and wait for description to be read. Made more variables global to make calling between functions easier (Phil Loy and Keisha Shepherd)
Added additional logging for debugging.
Made method that will wait for the exercise interval and break to start and finish before querying the next exercise.
Total Workout Time Left is it's own set interval function that can be started, paused, and cleared as necessary.
Created new method for starting exercise, timers, and music that is timed perfectly, after the voice announces the exercise and/or exercise description.
Formatting